Change package structure
