+++ src/main/java/personal/delivery/menu/Menu.java
@@ -1,10 +1,7 @@
-package personal.delivery.domain;
+package personal.delivery.menu;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 @Entity
 @Getter
